// angular/src/app/shared/models/configuration/InvAgent.ts
export class InvAgent {

    inv_agent_id: number;
    inv_agent_status: string;
    inv_agent_chk: number;
    inv_agent_name: string;
    inv_agent_shortname: string;
    inv_agent_type: string;
    inv_agent_extension: string;
    inv_agent_legal_id: string;
    inv_agent_internal_id: string;
    inv_agent_supervisor_id: string;
    inv_agent_supervisor_name: string;
    inv_agent_schedule_id: string;
    inv_agent_schedule_name: string;
    inv_agent_password: string;
    inv_agent_eccp_password: string;

}